Move tests to py3 file
